requirements.txt
aiogram==3.9.0
annotated-types==0.7.0
attrs==23.2.0
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==3.3.2
Deprecated==1.2.14
distro==1.9.0
frozenlist==1.4.1
google-auth==2.31.0
google-auth-oauthlib==1.2.0
gspread==6.1.2
httplib2==0.22.0
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
netaddr==1.3.0
numpy==2.0.0
oauth2client==4.1.3
oauthlib==3.2.2
pandas==2.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wrapt==1.16.0
yarl==1.9.4
yookassa==3.3.0
YooMoney==0.1.0
